+++ src/Twilio/Rest/Numbers/V1/PortingPortabilityOptions.php
@@ -22,16 +22,19 @@ abstract class PortingPortabilityOptions
 {
     /**
      * @param string $targetAccountSid Account Sid to which the number will be ported. This can be used to determine if a sub account already has the number in its inventory or a different sub account. If this is not provided, the authenticated account will be assumed to be the target account.
+     * @param string $addressSid Address Sid of customer to which the number will be ported.
      * @return FetchPortingPortabilityOptions Options builder
      */
     public static function fetch(
         
-        string $targetAccountSid = Values::NONE
+        string $targetAccountSid = Values::NONE,
+        string $addressSid = Values::NONE
 
     ): FetchPortingPortabilityOptions
     {
         return new FetchPortingPortabilityOptions(
-            $targetAccountSid
+            $targetAccountSid,
+            $addressSid
         );
     }
 
@@ -41,13 +44,16 @@ class FetchPortingPortabilityOptions extends Options
     {
     /**
      * @param string $targetAccountSid Account Sid to which the number will be ported. This can be used to determine if a sub account already has the number in its inventory or a different sub account. If this is not provided, the authenticated account will be assumed to be the target account.
+     * @param string $addressSid Address Sid of customer to which the number will be ported.
      */
     public function __construct(
         
-        string $targetAccountSid = Values::NONE
+        string $targetAccountSid = Values::NONE,
+        string $addressSid = Values::NONE
 
     ) {
         $this->options['targetAccountSid'] = $targetAccountSid;
+        $this->options['addressSid'] = $addressSid;
     }
 
     /**
@@ -59,6 +65,18 @@ class FetchPortingPortabilityOptions extends Options
     public function setTargetAccountSid(string $targetAccountSid): self
     {
         $this->options['targetAccountSid'] = $targetAccountSid;
+        return $this;
+    }
+
+    /**
+     * Address Sid of customer to which the number will be ported.
+     *
+     * @param string $addressSid Address Sid of customer to which the number will be ported.
+     * @return $this Fluent Builder
+     */
+    public function setAddressSid(string $addressSid): self
+    {
+        $this->options['addressSid'] = $addressSid;
         return $this;
     }