+++ src/Twilio/Rest/PreviewBase.php
@@ -14,21 +14,18 @@ namespace Twilio\Rest;
 
 use Twilio\Domain;
 use Twilio\Exceptions\TwilioException;
-use Twilio\Rest\Preview\DeployedDevices;
 use Twilio\Rest\Preview\HostedNumbers;
 use Twilio\Rest\Preview\Sync;
 use Twilio\Rest\Preview\Marketplace;
 use Twilio\Rest\Preview\Wireless;
 
 /**
- * @property \Twilio\Rest\Preview\DeployedDevices $deployedDevices
  * @property \Twilio\Rest\Preview\HostedNumbers $hostedNumbers
  * @property \Twilio\Rest\Preview\Sync $sync
  * @property \Twilio\Rest\Preview\Marketplace $marketplace
  * @property \Twilio\Rest\Preview\Wireless $wireless
  */
 class PreviewBase extends Domain {
-    protected $_deployedDevices;
     protected $_hostedNumbers;
     protected $_sync;
     protected $_marketplace;
@@ -45,16 +42,6 @@ class PreviewBase extends Domain {
         $this->baseUrl = 'https://preview.twilio.com';
     }
 
-
-    /**
-     * @return DeployedDevices Version deployedDevices of preview
-     */
-    protected function getDeployedDevices(): DeployedDevices {
-        if (!$this->_deployedDevices) {
-            $this->_deployedDevices = new DeployedDevices($this);
-        }
-        return $this->_deployedDevices;
-    }
 
     /**
      * @return HostedNumbers Version hostedNumbers of preview