+++ src/Twilio/Rest/Verify/V2/Service/VerificationList.php
@@ -94,6 +94,8 @@ class VerificationList extends ListResource
                 $options['templateCustomSubstitutions'],
             'DeviceIp' =>
                 $options['deviceIp'],
+            'EnableSnaClientToken' =>
+                Serialize::booleanToString($options['enableSnaClientToken']),
             'RiskCheck' =>
                 $options['riskCheck'],
             'Tags' =>